Refractoring for clarity
